Added new logic. Update controller. Update DAO and SERVICE layer.

+++ src/main/java/shop/domain/CarBrandModel.java
@@ -7,14 +7,15 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
 import javax.persistence.OneToOne;
 
 @Entity
 public class CarBrandModel {
+	
 	@Id
-	@GeneratedValue(strategy=GenerationType.IDENTITY) 
-	private int id;
+	@GeneratedValue(strategy = GenerationType.IDENTITY) 
+	private Long id;
+	
 	@Column(name = "name")
 	private String name;
 	
@@ -31,11 +32,11 @@ public class CarBrandModel {
 		this.name = name;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
