+++ src/main/java/shop/domain/CarTypeOfDrive.java
@@ -16,7 +16,8 @@ public class CarTypeOfDrive {
 	
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY) 
-	private int id;
+	private Long id;
+	
 	@Column(name = "name")
 	private String name;
 	
@@ -31,11 +32,11 @@ public class CarTypeOfDrive {
 		this.name = name;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
